add CategoryController change path categoryList.html, categoryCreate.html
